fix constant types c#

+++ crates/servicepoint_binding_cs/ServicePoint/Constants.cs
@@ -5,20 +5,20 @@ namespace ServicePoint;
 public static class Constants
 {
     /// size of a single tile in one dimension
-    public const int TileSize = NativeMethods.SP_TILE_SIZE;
+    public const nuint TileSize = NativeMethods.SP_TILE_SIZE;
 
     /// tile count in the x-direction
-    public const int TileWidth = NativeMethods.SP_TILE_WIDTH;
+    public const nuint TileWidth = NativeMethods.SP_TILE_WIDTH;
 
     /// tile count in the y-direction
-    public const int TileHeight = NativeMethods.SP_TILE_SIZE;
+    public const nuint TileHeight = NativeMethods.SP_TILE_SIZE;
 
     /// screen width in pixels
-    public const int PixelWidth = TileWidth * TileSize;
+    public const nuint PixelWidth = TileWidth * TileSize;
 
     /// screen height in pixels
-    public const int PixelHeight = TileHeight * TileSize;
+    public const nuint PixelHeight = TileHeight * TileSize;
 
     /// pixel count on whole screen
-    public const int PixelCount = PixelWidth * PixelHeight;
+    public const nuint PixelCount = PixelWidth * PixelHeight;
 }
